Create $avatarID, store avatar in registry
